Stock constructor and loading methods modified
