+++ src/Plate/index.ts
@@ -1,5 +1,5 @@
 export { Ingredient } from './ingredient';
-export { BasicPlate } from './basic_plate';
+export { BasicPlate, PlateType } from './basic_plate';
 export { StarterPlate } from './starter_plate';
 export { FirstPlate } from './first_plate';
 export { SecondPlate } from './second_plate';
