+++ tests/Stock/database_test.json
@@ -2,7 +2,7 @@
   "stock": {
     "foods": [],
     "plates": [],
-    "Menus": [],
-    "cartas": []
+    "cartas": [],
+    "menus": []
   }
 }
